fix: break out checkbox components
