+++ src/components/Checkbox/Checkbox.js
@@ -1,45 +1,18 @@
 import React from "react";
-import styled from "styled-components";
 import { PropTypes } from "prop-types";
-import { COLOR_BRAND_PRIMARY } from "style/styleVariables";
 
-const Box = styled.input.attrs({
-  type: "checkbox"
-})`
-  appearance: none;
-  background-color: #fafafa;
-  border: 1px solid ${COLOR_BRAND_PRIMARY};
-  box-shadow: 0 1px 2px rgba(0, 0, 0, 0.05),
-    inset 0px -15px 10px -12px rgba(0, 0, 0, 0.05);
-  padding: 6px;
-  height: 8px;
-  position: relative;
-  border-radius: 3px;
-  &:checked {
-    &:after {
-      content: "\u2714";
-      font-size: 10px;
-      position: absolute;
-      top: 0px;
-      left: 2px;
-      color: ${COLOR_BRAND_PRIMARY};
-    }
-  }
-`;
+import Label from "./Label";
+import Box from "./Box";
 
-const Label = styled.label`
-  display: flex;
-  align-items: center;
-  > input {
-    margin: 0 8px;
-  }
-`;
-
-//pick out label, all other props are passed to checkbox element
-const Checkbox = ({ label, defaultChecked, ...props }) => {
+const Checkbox = ({ label, value, defaultChecked, ...props }) => {
   return (
     <Label for={label}>
-      <Box defaultChecked={defaultChecked} {...props} />
+      <Box
+        defaultChecked={defaultChecked}
+        name={label}
+        value={value}
+        {...props}
+      />
       {label}
     </Label>
   );
@@ -48,7 +21,8 @@ const Checkbox = ({ label, defaultChecked, ...props }) => {
 Checkbox.propTypes = {
   defaultChecked: PropTypes.bool,
   label: PropTypes.string.isRequired,
-  onChange: PropTypes.func
+  onChange: PropTypes.func,
+  value: PropTypes.string
 };
 
 export default Checkbox;
