+++ src/components/Checkbox/components/Box.test.js
@@ -1,0 +1,12 @@
+import React from "react";
+import renderer from "react-test-renderer";
+import "jest-styled-components";
+
+import Box from "./Box";
+
+describe("Box", () => {
+  it("should render", () => {
+    const aBox = renderer.create(<Box />).toJSON();
+    expect(aBox).toMatchSnapshot();
+  });
+});
