+++ src/components/Checkbox/components/Label.js
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+
+const Label = styled.label`
+  display: flex;
+  align-items: center;
+  > input {
+    margin: 0 8px;
+  }
+`;
+
+export default Label;
